+++ src/index.js
@@ -1,134 +1,23 @@
-import express from "express";
-import { v4 as uuidv4 } from "uuid";
+import dotenv from "dotenv";
 import cors from "cors";
-import { ApolloServer, gql } from "apollo-server-express";
+import express from "express";
+import { ApolloServer } from "apollo-server-express";
 
+import schema from "./schema";
+import resolvers from "./resolvers";
+import models from "./models";
+
+dotenv.config();
 const app = express();
 app.use(cors());
 
-const schema = gql`
-  type Query {
-    users: [User!]
-    user(id: ID!): User
-    me: User
-
-    messages: [Message!]!
-    message(id: ID!): Message!
-  }
-
-  type Mutation {
-    createMessage(text: String!): Message!
-    deleteMessage(id: ID!): Boolean!
-  }
-
-  type User {
-    id: ID!
-    username: String!
-    messages: [Message!]
-  }
-
-  type Message {
-    id: ID!
-    text: String!
-    userId: ID!
-  }
-`;
-
-let users = {
-  1: {
-    id: "1",
-    username: "Robin Wieruch",
-    messageIds: [1],
-  },
-  2: {
-    id: "2",
-    username: "Dave Davids",
-    messageIds: [2],
-  },
-};
-
-const me = users[1];
-
-let messages = {
-  1: {
-    id: "1",
-    text: "Hello World",
-    userId: "1",
-  },
-  2: {
-    id: "2",
-    text: "By World",
-    userId: "2",
-  },
-};
-
-const resolvers = {
-  Query: {
-    users: () => {
-      return Object.values(users);
-    },
-    user: (parent, { id }) => {
-      return users[id];
-    },
-    me: (parent, args, { me }) => {
-      return me;
-    },
-    messages: () => {
-      return Object.values(messages);
-    },
-    message: (parent, { id }) => {
-      return messages[id];
-    },
-  },
-
-  Mutation: {
-    createMessage: (parent, { text }, { me }) => {
-      const id = uuidv4();
-      const message = {
-        id,
-        text,
-        userId: me.id,
-      };
-
-      messages[id] = message;
-      users[me.id].messageIds.push(id);
-
-      return message;
-    },
-
-    deleteMessage: (parent, { id }) => {
-      const { [id]: message, ...otherMessages } = messages;
-
-      if (!message) {
-        return false;
-      }
-
-      messages = otherMessages;
-
-      return true;
-    },
-  },
-
-  User: {
-    messages: (user) => {
-      return Object.values(messages).filter(
-        (message) => message.userId === user.id
-      );
-    },
-  },
-
-  Message: {
-    user: (message) => {
-      return users[message.userId];
-    },
-  },
-};
 
 const server = new ApolloServer({
   typeDefs: schema,
   resolvers,
   context: {
-    me: users[1],
+    models,
+    me: models.users[1],
   },
 });
