+++ src/resolvers/user.js
@@ -1,0 +1,21 @@
+export default {
+  Query: {
+    users: (parent, args, { models }) => {
+      return Object.values(models.users);
+    },
+    user: (parent, { id }, { models }) => {
+      return models.users[id];
+    },
+    me: (parent, args, { me }) => {
+      return me;
+    },
+  },
+ 
+  User: {
+    messages: (user, args, { models }) => {
+      return Object.values(models.messages).filter(
+        (message) => message.userId === user.id
+      );
+    },
+  },
+};
